programmers/src/test/java/com/study/programmers/dfs/OilDrillingTest.java
package com.study.programmers.dfs;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class OilDrillingTest {
    OilDrilling oilDrilling = new OilDrilling();

    @Test
    void test01() {
        assertThat(oilDrilling.solution(new int[][]{{0, 0, 0, 1, 1, 1, 0, 0}, {0, 0, 0, 0, 1, 1, 0, 0}, {1, 1, 0, 0, 0, 1, 1, 0}, {1, 1, 1, 0, 0, 0, 0, 0}, {1, 1, 1, 0, 0, 0, 1, 1}}))
                .isEqualTo(9);
    }

    @Test
    void test02() {
        assertThat(oilDrilling.solution(new int[][]{{1, 0, 1, 0, 1, 1}, {1, 0, 1, 0, 0, 0}, {1, 0, 1, 0, 0, 1}, {1, 0, 0, 1, 0, 0}, {1, 0, 0, 1, 0, 1}, {1, 0, 0, 0, 0, 0}, {1, 1, 1, 1, 1, 1}}))
                .isEqualTo(16);
    }
}
